requirements.txt
datasets==3.5.1
matplotlib==3.10.7
numpy==2.3.5
PyYAML==6.0.3
scikit_learn==1.7.2
scipy==1.16.3
torch==2.9.1
